Differentiate accessor creation based on attribute's constness
For this purpose, we add a new overload taking only a getter as the
argument constructing a method accessor with a "null" setter. That
overload is only enabled if the attribute's type is const.

The other overloads are only enabled if the attribute is non-const. The
default-attribute was removed for the overloads, enforcing that the user
supplies a setter for non-const accessors.