Add SFINAE helper for detecting attribute accessors
With constructors, we have accessors which do not correspond to an
attribute. Hence, we can not rely on a member `attr` to be present.

This commit adds a SFINAE helber which detects the presence/absence of
this member and checks whether the member is identical to a specific
attribute.

The utility will be used for enabling and disabling getters and setters
in the accessor bundle.