Rewrite utility for selecting attribute values from parameters
It turned out the utility malfunctioned because of template parameters
which were not forwarded correctly. Attempts to fix the issue revealed
some fundamental flaws in the overloading approach.

Now we make use of `std::enable_if` to select the correct overload. We
also improved readability by introducing a little helper template.