Move cmoh::attribute::method_accessor to new namespace
We intend to introduce new attribute accessor types in the near future.
Putting them within the declaration of `attribute` would lead to a long
and potentially riddled declaration. Also, having each accessor in a
seperate file would be more modular.

Initially, the major reason why we even started putting the method
accessor within `cmoh::attribute` were of technical nature. However,
those reasons no longer exist. Hence, there is not good reason for
keeping the thing encapsulated in the attribute.

On the other hand, we don't want a bunch of accessors lying around
directly in the `cmoh` namespace. We therefore add a new namespace for
accessors, which in turn contains a namespace for attribute accessors,
knowing that we will eventually add method accessors.