Split by_method attribute accessor in two parts
One is an accessor that allows only queries of attributes, the second
one extends the first one by setting functionality. This change prevents
the user from setting a const attribute, because the accessor is no
longer able to do it.

On top of that, it strips a few bytes from the memory requirement of an
accessor for const attributes.